:art: edit diary header and delete unused files
