+++ app/(tabs)/_layout.tsx
@@ -8,12 +8,10 @@ import {
   MessageCircleMore,
 } from "lucide-react-native";
 import { useRouter } from "expo-router";
-import { ButtonComponents } from "@/components/Buntton";
-import { useCompare } from "@/context/CompareContext";
 
 export default function _layout() {
   const router = useRouter();
-  const { isCompare, setIsCompare, setCompare } = useCompare();
+
 
   return (
     <Tabs
@@ -39,28 +37,9 @@ export default function _layout() {
       <Tabs.Screen
         name="diary"
         options={{
-          headerTitle: isCompare ? "" : "Diary",
-          headerShown: true,
           tabBarIcon: ({ color, size }) => (
             <BookAIcon size={size} color={color} />
           ),
-          headerRight: () =>
-            isCompare ? <ConfirmButton /> : <CompareButton />,
-          headerLeft: () =>
-            isCompare ? (
-              <ButtonComponents
-                title="Cancel"
-                textSize="text-label8 font-semibold"
-                onPress={() => {
-                  setIsCompare(!isCompare);
-                  setCompare([]);
-                }}
-                className="px-2 py-2"
-              />
-            ) : null,
-          headerStyle: {
-            backgroundColor: "#FCFAFD",
-          },
         }}
       />
       <Tabs.Screen
@@ -86,31 +65,3 @@ export default function _layout() {
     </Tabs>
   );
 }
-
-const CompareButton = () => {
-  const { setIsCompare, isCompare } = useCompare();
-  return (
-    <ButtonComponents
-      title="Compare"
-      className="bg-Bittersweet px-2 py-2 rounded-full"
-      textSize="text-label8 font-semibold text-white"
-      onPress={() => setIsCompare(!isCompare)}
-    />
-  );
-};
-
-const ConfirmButton = () => {
-  const { setIsCompare, isCompare } = useCompare();
-  const router = useRouter();
-  return (
-    <ButtonComponents
-      title="Confirm"
-      className="bg-Bittersweet px-2 py-2 rounded-full"
-      textSize="text-label8 font-semibold text-white"
-      onPress={() => {
-        setIsCompare(!isCompare);
-        router.push("/compare");
-      }}
-    />
-  );
-};
