+++ app/(tabs)/diary/_layout.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { Stack } from "expo-router";
+
+export default function DiaryLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    >
+      <Stack.Screen name="index" />
+    </Stack>
+  );
+}
